add cache invalidation

+++ PlaceBot/Pixel.cs
@@ -2,11 +2,11 @@ namespace PlaceBot;
 
 public readonly struct Pixel
 {
-    public int X { get; }
-    public int Y { get; }
+    public ushort X { get; }
+    public ushort Y { get; }
     public byte Color { get; }
 
-    public Pixel(int x, int y, byte color)
+    public Pixel(ushort x, ushort y, byte color)
     {
         X = x;
         Y = y;
